ExposeJWKS, Migration, CustomJWKS and added my own Identity Providers
I have talked to Liam about this already but I was really blocked the way you update `auth`, my env was giving tons of trouble. I also experience alot bugs while pushing to the cloud. I am still working on how to intergate the change of sign-in flow just for my env. I have made comments as you can see in config-overide.js and in app.js. I also still looking how to make the auth record the trigger of the customjwks. Temporany I had just added the function trigger using console to not be a block and just develop the function. I will figure all this out in the next PR.